requirements.txt
langchain
langchain-google-genai
python-dotenv
pydantic
langchain-community 
chromadb
tavily-python 
beautifulsoup4
pypdf
fastapi